+++ Saturn/src/Saturn/Core/Random.h
@@ -35,7 +35,7 @@ namespace Saturn {
 	public:
 		static bool RandomBool();
 		static uint64_t RandomUUID();
-		static size_t RandomElementInRange( uint64_t min, uint64_t max );
+		static size_t RandomElementInRange( size_t min, size_t max );
 
 		static float RandomFloatInRange( float min, float max );
 	};
